Post Alt HCP UserEval Code

+++ include/dealprotos.h
@@ -43,7 +43,7 @@ extern void              setshapebit (int cl, int di, int ht, int sp, int msk, i
 extern void              clearpointcount ();                 /* yyparse action clause set 13 elem array to zero*/
 extern void              clearpointcount_alt (int cin);      /* yyparse action clause affects alternate counts */
 extern void              pointcount (int index, int value);  /* yyparse action clause affects HCPs*/
-extern void              zerocount (int points[13]);         /* called by clearpointcount etc. */
+extern void              zerocount (short int points[13]);   /* called by clearpointcount etc. */
 extern struct tree       *newquery(int tag, int side, int compass, int suit, int idx); /* yyparse action clause */
 extern char              *newpar_cstr(int vuln) ;           /* yyparse action clause. see dealparse_subs */
 extern struct csvterm_st *new_csvterm (struct tree *tr1, char *str1, int hand_mask,  int trix_mask, int par_mask, struct csvterm_st *csv1);
